Chg: Maps updated - Acropolis, Cobalt, Disco Bloodbath, Thunderbird, and Winter's Gate removed; Deathaura, Ice and Chrom, Pillars of Gold, and Submarine added.

+++ scctool/view/widgets.py
@@ -191,7 +191,7 @@ class MapDownloader(QProgressDialog):
         self.setValue(self.minimum())
 
         self.resize(QSize(
-            mainWindow.size().width() * 0.8, self.sizeHint().height()))
+            int(mainWindow.size().width() * 0.8), self.sizeHint().height()))
         relativeChange = QPoint(int(mainWindow.size().width() / 2),
                                 int(mainWindow.size().height() / 3))\
             - QPoint(int(self.size().width() / 2),
